refactor(ui5-flexible-column-layout): migrate to TypeScript
