+++ packages/fiori/src/fcl-utils/FCLLayout.ts
@@ -1,22 +1,37 @@
-const getLayoutsByMedia = () => {
+import { MEDIA } from "../FlexibleColumnLayout.js";
+
+type LayoutConfiguration = {
+	[key in MEDIA]: {
+		[key: string]: {
+			layout: Array<string>;
+			arrows: Array<{
+				visible: boolean;
+				dir: null | string;
+				separator?: boolean;
+			}>;
+		};
+	};
+};
+
+const getLayoutsByMedia = (): LayoutConfiguration => {
 	return {
 		desktop: {
 			"OneColumn": {
-				layout: ["100%", 0, 0],
+				layout: ["100%", "0px", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"TwoColumnsStartExpanded": {
-				layout: ["67%", "33%", 0],
+				layout: ["67%", "33%", "0px"],
 				arrows: [
 					{ visible: true, dir: "mirror" },
 					{ visible: false, dir: null },
 				],
 			},
 			"TwoColumnsMidExpanded": {
-				layout: ["33%", "67%", 0],
+				layout: ["33%", "67%", "0px"],
 				arrows: [
 					{ visible: true, dir: null },
 					{ visible: false, dir: null },
@@ -44,28 +59,28 @@ const getLayoutsByMedia = () => {
 				],
 			},
 			"ThreeColumnsStartExpandedEndHidden": {
-				layout: ["67%", "33%", 0],
+				layout: ["67%", "33%", "0px"],
 				arrows: [
 					{ visible: true, dir: "mirror" },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsMidExpandedEndHidden": {
-				layout: ["33%", "67%", 0],
+				layout: ["33%", "67%", "0px"],
 				arrows: [
 					{ visible: true, dir: null },
 					{ visible: true, dir: null },
 				],
 			},
 			"MidColumnFullScreen": {
-				layout: [0, "100%", 0],
+				layout: ["0px", "100%", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"EndColumnFullScreen": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
@@ -74,70 +89,70 @@ const getLayoutsByMedia = () => {
 		},
 		tablet: {
 			"OneColumn": {
-				layout: ["100%", 0, 0],
+				layout: ["100%", "0px", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"TwoColumnsStartExpanded": {
-				layout: ["67%", "33%", 0],
+				layout: ["67%", "33%", "0px"],
 				arrows: [
 					{ visible: true, dir: "mirror" },
 					{ visible: false, dir: null },
 				],
 			},
 			"TwoColumnsMidExpanded": {
-				layout: ["33%", "67%", 0],
+				layout: ["33%", "67%", "0px"],
 				arrows: [
 					{ visible: true, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsStartExpanded": {
-				layout: ["67%", "33%", 0],
+				layout: ["67%", "33%", "0px"],
 				arrows: [
 					{ visible: true, dir: "mirror" },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsMidExpanded": {
-				layout: [0, "67%", "33%"],
+				layout: ["0px", "67%", "33%"],
 				arrows: [
 					{ visible: true, dir: null },
 					{ visible: true, dir: null },
 				],
 			},
 			"ThreeColumnsEndExpanded": {
-				layout: [0, "33%", "67%"],
+				layout: ["0px", "33%", "67%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: true, dir: "mirror" },
 				],
 			},
 			"ThreeColumnsStartExpandedEndHidden": {
-				layout: ["67%", "33%", 0],
+				layout: ["67%", "33%", "0px"],
 				arrows: [
 					{ visible: true, dir: "mirror" },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsMidExpandedEndHidden": {
-				layout: ["33%", "67%", 0],
+				layout: ["33%", "67%", "0px"],
 				arrows: [
 					{ visible: true, dir: null },
 					{ visible: true, dir: null },
 				],
 			},
 			"MidColumnFullScreen": {
-				layout: [0, "100%", 0],
+				layout: ["0px", "100%", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"EndColumnFullScreen": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
@@ -146,70 +161,70 @@ const getLayoutsByMedia = () => {
 		},
 		phone: {
 			"OneColumn": {
-				layout: ["100%", 0, 0],
+				layout: ["100%", "0px", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"TwoColumnsStartExpanded": {
-				layout: [0, "100%", 0],
+				layout: ["0px", "100%", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"TwoColumnsMidExpanded": {
-				layout: [0, "100%", 0],
+				layout: ["0px", "100%", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsStartExpanded": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsMidExpanded": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsEndExpanded": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsStartExpandedEndHidden": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"ThreeColumnsMidExpandedEndHidden": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"MidColumnFullScreen": {
-				layout: [0, "100%", 0],
+				layout: ["0px", "100%", "0px"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
 				],
 			},
 			"EndColumnFullScreen": {
-				layout: [0, 0, "100%"],
+				layout: ["0px", "0px", "100%"],
 				arrows: [
 					{ visible: false, dir: null },
 					{ visible: false, dir: null },
@@ -243,4 +258,8 @@ export {
 	getLayoutsByMedia,
 	getNextLayoutByStartArrow,
 	getNextLayoutByEndArrow,
+};
+
+export type {
+	LayoutConfiguration,
 };
