+++ packages/fiori/test/pages/FCLCustom.html
@@ -70,21 +70,21 @@
 		fcl._layoutsConfiguration = {
 			desktop: {
 				"OneColumn": {
-					layout: ["100%", 0, 0],
+					layout: ["100%", "0px", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"TwoColumnsStartExpanded": {
-					layout: ["90%", "10%", 0],
+					layout: ["90%", "10%", "0px"],
 					arrows: [
 						{ visible: true, dir: "mirror" },
 						{ visible: false, dir: null },
 					],
 				},
 				"TwoColumnsMidExpanded": {
-					layout: ["10%", "90%", 0],
+					layout: ["10%", "90%", "0px"],
 					arrows: [
 						{ visible: true, dir: null },
 						{ visible: false, dir: null },
@@ -112,28 +112,28 @@
 					],
 				},
 				"ThreeColumnsStartExpandedEndHidden": {
-					layout: ["90%", "10%", 0],
+					layout: ["90%", "10%", "0px"],
 					arrows: [
 						{ visible: true, dir: "mirror" },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsMidExpandedEndHidden": {
-					layout: ["10%", "90%", 0],
+					layout: ["10%", "90%", "0px"],
 					arrows: [
 						{ visible: true, dir: null },
 						{ visible: true, dir: null },
 					],
 				},
 				"MidColumnFullScreen": {
-					layout: [0, "100%", 0],
+					layout: ["0px", "100%", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"EndColumnFullScreen": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
@@ -142,70 +142,70 @@
 			},
 			tablet: {
 				"OneColumn": {
-					layout: ["100%", 0, 0],
+					layout: ["100%", "0px", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"TwoColumnsStartExpanded": {
-					layout: ["90%", "10%", 0],
+					layout: ["90%", "10%", "0px"],
 					arrows: [
 						{ visible: true, dir: "mirror" },
 						{ visible: false, dir: null },
 					],
 				},
 				"TwoColumnsMidExpanded": {
-					layout: ["10%", "90%", 0],
+					layout: ["10%", "90%", "0px"],
 					arrows: [
 						{ visible: true, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsStartExpanded": {
-					layout: ["90%", "10%", 0],
+					layout: ["90%", "10%", "0px"],
 					arrows: [
 						{ visible: true, dir: "mirror" },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsMidExpanded": {
-					layout: [0, "90%", "10%"],
+					layout: ["0px", "90%", "10%"],
 					arrows: [
 						{ visible: true, dir: null },
 						{ visible: true, dir: null },
 					],
 				},
 				"ThreeColumnsEndExpanded": {
-					layout: [0, "10%", "90%"],
+					layout: ["0px", "10%", "90%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: true, dir: "mirror" },
 					],
 				},
 				"ThreeColumnsStartExpandedEndHidden": {
-					layout: ["90%", "10%", 0],
+					layout: ["90%", "10%", "0px"],
 					arrows: [
 						{ visible: true, dir: "mirror" },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsMidExpandedEndHidden": {
-					layout: ["10%", "90%", 0],
+					layout: ["10%", "90%", "0px"],
 					arrows: [
 						{ visible: true, dir: null },
 						{ visible: true, dir: null },
 					],
 				},
 				"MidColumnFullScreen": {
-					layout: [0, "100%", 0],
+					layout: ["0px", "100%", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"EndColumnFullScreen": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
@@ -214,70 +214,70 @@
 			},
 			phone: {
 				"OneColumn": {
-					layout: ["100%", 0, 0],
+					layout: ["100%", "0px", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"TwoColumnsStartExpanded": {
-					layout: [0, "100%", 0],
+					layout: ["0px", "100%", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"TwoColumnsMidExpanded": {
-					layout: [0, "100%", 0],
+					layout: ["0px", "100%", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsStartExpanded": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsMidExpanded": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsEndExpanded": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsStartExpandedEndHidden": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"ThreeColumnsMidExpandedEndHidden": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"MidColumnFullScreen": {
-					layout: [0, "100%", 0],
+					layout: ["0px", "100%", "0px"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
 					],
 				},
 				"EndColumnFullScreen": {
-					layout: [0, 0, "100%"],
+					layout: ["0px", "0px", "100%"],
 					arrows: [
 						{ visible: false, dir: null },
 						{ visible: false, dir: null },
